refactor: 게시글 신고 기능 하나의 객체로 분리
