+++ src/main/java/dandi/dandi/post/application/port/in/PostReportUseCaseServicePort.java
@@ -1,0 +1,6 @@
+package dandi.dandi.post.application.port.in;
+
+public interface PostReportUseCaseServicePort {
+
+    void reportPost(Long memberId, Long postId);
+}
