+++ src/main/java/dandi/dandi/post/application/port/in/PostUseCaseServicePort.java
@@ -1,10 +1,8 @@
 package dandi.dandi.post.application.port.in;
 
-public interface PostUseCase {
+public interface PostUseCaseServicePort {
 
     PostRegisterResponse registerPost(Long memberId, PostRegisterCommand postRegisterCommand);
 
     void deletePost(Long memberId, Long postId);
-
-    void reportPost(Long memberId, Long postId);
 }
