+++ src/main/java/dandi/dandi/post/application/service/PostUseCaseServiceAdapter.java
@@ -4,9 +4,8 @@ import dandi.dandi.common.exception.ForbiddenException;
 import dandi.dandi.common.exception.NotFoundException;
 import dandi.dandi.post.application.port.in.PostRegisterCommand;
 import dandi.dandi.post.application.port.in.PostRegisterResponse;
-import dandi.dandi.post.application.port.in.PostUseCase;
+import dandi.dandi.post.application.port.in.PostUseCaseServicePort;
 import dandi.dandi.post.application.port.out.PostPersistencePort;
-import dandi.dandi.post.application.port.out.PostReportPersistencePort;
 import dandi.dandi.post.domain.Post;
 import dandi.dandi.post.domain.Temperatures;
 import dandi.dandi.post.domain.WeatherFeeling;
@@ -15,18 +14,16 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
-public class PostService implements PostUseCase {
+public class PostUseCaseServiceAdapter implements PostUseCaseServicePort {
 
     private static final int POST_IMAGE_URL_INDEX = 1;
 
     private final PostPersistencePort postPersistencePort;
-    private final PostReportPersistencePort postReportPersistencePort;
     private final String imageAccessUrl;
 
-    public PostService(PostPersistencePort postPersistencePort, PostReportPersistencePort postReportPersistencePort,
-                       @Value("${cloud.aws.cloud-front.uri}") String imageAccessUrl) {
+    public PostUseCaseServiceAdapter(PostPersistencePort postPersistencePort,
+                                     @Value("${cloud.aws.cloud-front.uri}") String imageAccessUrl) {
         this.postPersistencePort = postPersistencePort;
-        this.postReportPersistencePort = postReportPersistencePort;
         this.imageAccessUrl = imageAccessUrl;
     }
 
@@ -60,26 +57,6 @@ public class PostService implements PostUseCase {
     public void validateDeleteAuthorization(Post post, Long memberId) {
         if (!post.isWrittenBy(memberId)) {
             throw ForbiddenException.postDeletion();
-        }
-    }
-
-    @Override
-    @Transactional
-    public void reportPost(Long memberId, Long postId) {
-        validatePostExistence(postId);
-        validateAlreadyReported(memberId, postId);
-        postReportPersistencePort.savePostReportOf(memberId, postId);
-    }
-
-    private void validatePostExistence(Long postId) {
-        if (!postPersistencePort.existsById(postId)) {
-            throw NotFoundException.post();
-        }
-    }
-
-    private void validateAlreadyReported(Long memberId, Long postId) {
-        if (postReportPersistencePort.existsByMemberIdAndPostId(memberId, postId)) {
-            throw new IllegalStateException("이미 신고한 게시글입니다.");
         }
     }
 }
