+++ src/test/java/dandi/dandi/post/application/service/PostUseCaseServiceAdapterTest.java
@@ -6,7 +6,6 @@ import static dandi.dandi.post.PostFixture.MAX_TEMPERATURE;
 import static dandi.dandi.post.PostFixture.MIN_TEMPERATURE;
 import static dandi.dandi.post.PostFixture.OUTFIT_FEELING_INDEX;
 import static dandi.dandi.post.PostFixture.POST;
-import static dandi.dandi.post.PostFixture.POST_ID;
 import static dandi.dandi.post.PostFixture.POST_IMAGE_FULL_URL;
 import static dandi.dandi.utils.TestImageUtils.IMAGE_ACCESS_URL;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThat;
@@ -16,12 +15,9 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import dandi.dandi.common.exception.ForbiddenException;
-import dandi.dandi.common.exception.NotFoundException;
 import dandi.dandi.post.application.port.in.PostRegisterCommand;
 import dandi.dandi.post.application.port.out.PostPersistencePort;
-import dandi.dandi.post.application.port.out.PostReportPersistencePort;
 import dandi.dandi.post.domain.Post;
-import dandi.dandi.postlike.application.port.out.PostLikePersistencePort;
 import java.util.Optional;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -30,13 +26,11 @@ import org.mockito.Mockito;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 @ExtendWith(MockitoExtension.class)
-class PostServiceTest {
+class PostUseCaseServiceAdapterTest {
 
     private final PostPersistencePort postPersistencePort = Mockito.mock(PostPersistencePort.class);
-    private final PostLikePersistencePort postLikePersistencePort = Mockito.mock(PostLikePersistencePort.class);
-    private final PostReportPersistencePort postReportPersistencePort = Mockito.mock(PostReportPersistencePort.class);
-    private final PostService postService = new PostService(postPersistencePort, postReportPersistencePort,
-            IMAGE_ACCESS_URL);
+    private final PostUseCaseServiceAdapter postService =
+            new PostUseCaseServiceAdapter(postPersistencePort, IMAGE_ACCESS_URL);
 
     @DisplayName("게시글을 작성할 수 있다.")
     @Test
@@ -76,42 +70,5 @@ class PostServiceTest {
         assertThatThrownBy(() -> postService.deletePost(postDeletionForbiddenMemberId, postId))
                 .isInstanceOf(ForbiddenException.class)
                 .hasMessage(ForbiddenException.postDeletion().getMessage());
-    }
-
-    @DisplayName("존재하지 않는 글을 신고하려고 하면 예외를 발생시킨다.")
-    @Test
-    void reportPost_NotFound() {
-        when(postPersistencePort.existsById(POST_ID))
-                .thenReturn(false);
-
-        assertThatThrownBy(() -> postService.reportPost(MEMBER_ID, POST_ID))
-                .isInstanceOf(NotFoundException.class)
-                .hasMessage(NotFoundException.post().getMessage());
-    }
-
-    @DisplayName("이미 신고한 글을 신고하려고 하면 예외를 발생시킨다.")
-    @Test
-    void reportPost_AlreadyReported() {
-        when(postPersistencePort.existsById(POST_ID))
-                .thenReturn(true);
-        when(postReportPersistencePort.existsByMemberIdAndPostId(MEMBER_ID, POST_ID))
-                .thenReturn(true);
-
-        assertThatThrownBy(() -> postService.reportPost(MEMBER_ID, POST_ID))
-                .isInstanceOf(IllegalStateException.class)
-                .hasMessage("이미 신고한 게시글입니다.");
-    }
-
-    @DisplayName("게시글을 신고할 수 있다.")
-    @Test
-    void reportPost() {
-        when(postPersistencePort.existsById(POST_ID))
-                .thenReturn(true);
-        when(postReportPersistencePort.existsByMemberIdAndPostId(MEMBER_ID, POST_ID))
-                .thenReturn(false);
-
-        postService.reportPost(MEMBER_ID, POST_ID);
-
-        verify(postReportPersistencePort).savePostReportOf(MEMBER_ID, POST_ID);
     }
 }
